Fix memory leak in glyph cache of server. Fix memory leak in client.

+++ gdipp_demo/demo_painter.h
@@ -15,7 +15,7 @@ public:
 	void stop_painting();
 
 private:
-	int _demo_font_count;
+	size_t _demo_font_count;
 	const wchar_t **_demo_fonts;
 	bool _random_text;
 	int _total_cycles;
